+++ src/pypnnomenclature/migrations/data/data_nomenclatures_taxonomie.sql
@@ -6,7 +6,7 @@ SET check_function_bodies = false;
 SET client_min_messages = warning;
 --SET row_security = off;
 
-SET search_path = ref_nomenclatures, pg_catalog;
+SET search_path = ref_nomenclatures, pg_catalog, public;
 
 
 TRUNCATE TABLE cor_taxref_nomenclature;